+++ src/scripts/components/utils/index.js
@@ -1,0 +1,1 @@
+export { default, isWeekend,getDaysInMonth } from "./utils";
