+++ src/styles/index.scss
@@ -1,3 +1,4 @@
+
 html,
 body {
   font-family: "Open Sans", sans-serif;
@@ -40,6 +41,7 @@ table {
   border: 1px solid rgba(126, 126, 126, 0.3);
   width: 34px;
   height: 33px;
+  transition: background-color 0.3s;
 
   &:hover {
     background: rgba(155, 153, 255, 0.1);
@@ -77,6 +79,7 @@ table {
     top: 0;
     left: 0;
     cursor: pointer;
+    transition: background-color 0.3s;
 
     &:hover {
       background: #fff;
